+++ CHANGELOG.md
@@ -16,6 +16,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - react-scripts package
 
+## [1.0.1-beta.2] - 2019-09-11
+- Package Dependency is updated.(#310)
+
+## [1.0.1-beta.1] (Resumption Phase) - 2019-09-04
+- Resuming and completing gauge component
+- Adding the Lattice-SSR example.
+
 ## [1.0.1-beta.1] - 2018-11-13 [YANKED]
 
 ## 1.0.1-beta.0 - 2018-11-08